+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- Types: `BooleanType`, `DataType`, `DataTypes`, `FloatType`, `IntegerType`, `LongType`, `StringType`, `StructField`, `StructType`
 
 <small>(formatted per [keepachangelog-1.1.0](http://keepachangelog.com/en/1.0.0/))</small>